Add `RawArrayVec` backing both `ArrayString` and `ArrayVec`
`RawArrayVec` does not concern itself with the dropping or non-dropping
of elements when it goes out of scope.
