+++ src/string.rs
@@ -5,11 +5,10 @@ use std::mem;
 use std::ptr;
 use std::ops::{Deref, DerefMut};
 use std::str;
-use std::slice;
 
-use array::Array;
-use array::Index;
+use Array;
 use CapacityError;
+use raw::RawArrayVec;
 
 /// A string with a fixed capacity.
 ///
@@ -18,11 +17,11 @@ use CapacityError;
 ///
 /// The string is a contiguous value that you can store directly on the stack
 /// if needed.
-#[derive(Copy)]
 pub struct ArrayString<A: Array<Item=u8>> {
-    xs: A,
-    len: A::Index,
+    inner: RawArrayVec<A>,
 }
+
+impl<A: Array<Item=u8> + Copy> Copy for ArrayString<A> where A::Item: Clone { }
 
 impl<A: Array<Item=u8>> ArrayString<A> {
     /// Create a new empty `ArrayString`.
@@ -38,11 +37,8 @@ impl<A: Array<Item=u8>> ArrayString<A> {
     /// assert_eq!(string.capacity(), 16);
     /// ```
     pub fn new() -> ArrayString<A> {
-        unsafe {
-            ArrayString {
-                xs: ::new_array(),
-                len: Index::from(0),
-            }
+        ArrayString {
+            inner: RawArrayVec::new(),
         }
     }
 
@@ -75,7 +71,7 @@ impl<A: Array<Item=u8>> ArrayString<A> {
     /// assert_eq!(string.capacity(), 3);
     /// ```
     #[inline]
-    pub fn capacity(&self) -> usize { A::capacity() }
+    pub fn capacity(&self) -> usize { self.inner.capacity() }
 
     /// Return if the `ArrayString` is completely filled.
     ///
@@ -137,7 +133,7 @@ impl<A: Array<Item=u8>> ArrayString<A> {
             return Err(CapacityError::new(s));
         }
         unsafe {
-            let dst = self.xs.as_mut_ptr().offset(self.len() as isize);
+            let dst = self.inner.as_mut_ptr().offset(self.len() as isize);
             let src = s.as_ptr();
             ptr::copy_nonoverlapping(src, dst, s.len());
             let newl = self.len() + s.len();
@@ -161,8 +157,7 @@ impl<A: Array<Item=u8>> ArrayString<A> {
     /// number of “valid” bytes in the string. Use with care.
     #[inline]
     pub unsafe fn set_len(&mut self, length: usize) {
-        debug_assert!(length <= self.capacity());
-        self.len = Index::from(length);
+        self.inner.set_len(length)
     }
 
     /// Return a string slice of the whole `ArrayString`.
@@ -176,8 +171,7 @@ impl<A: Array<Item=u8>> Deref for ArrayString<A> {
     #[inline]
     fn deref(&self) -> &str {
         unsafe {
-            let sl = slice::from_raw_parts(self.xs.as_ptr(), self.len.to_usize());
-            str::from_utf8_unchecked(sl)
+            str::from_utf8_unchecked(&self.inner)
         }
     }
 }
@@ -186,9 +180,8 @@ impl<A: Array<Item=u8>> DerefMut for ArrayString<A> {
     #[inline]
     fn deref_mut(&mut self) -> &mut str {
         unsafe {
-            let sl = slice::from_raw_parts_mut(self.xs.as_mut_ptr(), self.len.to_usize());
-            // FIXME: Nothing but transmute to do this right now
-            mem::transmute(sl)
+            let bytes: &mut [u8] = &mut self.inner;
+            mem::transmute(bytes)
         }
     }
 }
@@ -246,10 +239,9 @@ impl<A: Array<Item=u8> + Copy> Clone for ArrayString<A> {
     fn clone(&self) -> ArrayString<A> {
         *self
     }
-
     fn clone_from(&mut self, rhs: &Self) {
         // guaranteed to fit due to types matching.
         self.clear();
-        self.push_str(rhs).ok();
+        let _ = self.push_str(rhs);
     }
 }
